Better APK downloading
Apk variants now appear as a list of possible
options, each of which downloads that specific
apk.
Also, made some changes to the app internal
update mechanism
