+++ app/src/main/java/news/androidtv/tvapprepo/model/Apk.java
@@ -1,7 +1,12 @@
 package news.androidtv.tvapprepo.model;
 
+import android.util.Log;
+
 import org.json.JSONException;
 import org.json.JSONObject;
+
+import java.util.HashMap;
+import java.util.Map;
 
 /**
  * Represents an app file.
@@ -10,8 +15,12 @@ import org.json.JSONObject;
  * @version 2016.09.08
  */
 public class Apk {
+    private static final String TAG = Apk.class.getSimpleName();
+    private static final boolean DEBUG = true;
+
     private String banner;
-    private String downloadUrl;
+    private Map<String, String> downloadUrl;
+//    private FirebaseMap downloadMap;
     private String icon;
     private String name;
     private String packageName;
@@ -29,8 +38,32 @@ public class Apk {
         return banner;
     }
 
-    public String getDownloadUrl() {
+    public Map<String, String> getDownloadUrl() {
         return downloadUrl;
+    }
+
+    private HashMap<String, String> getDownloadMap() {
+        /*if (downloadMap == null) {
+            downloadMap = new FirebaseMap(downloadUrl.toString());
+        }
+        return downloadMap.getMap();*/
+        return (HashMap<String, String>) downloadUrl;
+    }
+
+    public int getDownloadCount() {
+        return getDownloadMap().size();
+    }
+
+    public String getDefaultDownloadUrl() {
+        return getDownloadMap().get(getDownloadMap().keySet().toArray()[0]);
+    }
+
+    public CharSequence[] getDownloadTitleArray() {
+        return getDownloadMap().keySet().toArray(new String[getDownloadMap().keySet().size()]);
+    }
+
+    public String[] getDownloadUrlArray() {
+        return getDownloadMap().values().toArray(new String[getDownloadMap().keySet().size()]);
     }
 
     public String getIcon() {
@@ -105,13 +138,17 @@ public class Apk {
         }
 
         public Builder(String serial) {
+            Log.d(TAG, serial);
             try {
                 JSONObject jsonObject = new JSONObject(serial);
+                Log.d(TAG, jsonObject.toString());
                 mApk = new Apk();
                 mApk.name = jsonObject.getString("name");
                 mApk.banner = jsonObject.getString("banner");
                 mApk.icon = jsonObject.getString("icon");
-                mApk.downloadUrl = jsonObject.getString("downloadUrl");
+                String downloadUrl = jsonObject.getString("downloadUrl");
+                Log.d(TAG, downloadUrl);
+                mApk.downloadUrl = new FirebaseMap(downloadUrl).getMap();
                 mApk.packageName = jsonObject.getString("packageName");
                 mApk.submitted = jsonObject.getLong("submissionDate");
                 mApk.versionCode = jsonObject.getInt("versionCode");
