+++ app/src/main/res/layout/activity_details.xml
@@ -2,7 +2,7 @@
 <fragment xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/details_fragment"
-    android:name="news.androidtv.tvapprepo.fragments.VideoDetailsFragment"
+    android:name="news.androidtv.tvapprepo.fragments.AppDetailsFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     tools:context=".activities.DetailsActivity"
